requirements.txt
python-telegram-bot
web3
python-dotenv
google-generativeai